fix(windows): Call purs.cmd on windows if avail
On Windows, `purs` commands fail when purescript is installed through
npm. This is because Turtle commands can't seem to find the purs script
on the path (it isn't executable).

The work around up until now was to put the purs.exe on the path. This
change will look for the purs.cmd file on the path and call that if it
is available.

To make calling purs.cmd work, I had to change the Turtle calls from
using procs to using shells. This is because the `procs` calls cause the
`~%dp0` macros in the command file to expand from the "working"
directory, rather then from directory where the cmd file is located.